group compiler syntax modules
