+++ src/stdlib.rs
@@ -1,8 +1,8 @@
 use std::path::PathBuf;
 
-use crate::ast::Module;
+use crate::compiler::syntax::Module;
+use crate::compiler::syntax::parser::parse_module;
 use crate::error::{ArgentError, Result};
-use crate::parser::parse_module;
 
 pub const CORE_MODULE: &str = "std::core";
